+++ packages/core/src/m3/deserializer.ts
@@ -3,7 +3,7 @@ import { deserializeSerializationChunk } from "../deserializer.js"
 import { nodesExtractorUsing } from "../extraction.js"
 import { defaultSimplisticHandler, SimplisticHandler } from "../handler.js"
 import { BuiltinPropertyValueDeserializer, lioncoreBuiltins } from "./builtins.js"
-import { lioncoreReader, lioncoreInstantiationFacade } from "./facade.js"
+import { lioncoreReader, lioncoreWriter } from "./facade.js"
 import { lioncore } from "./lioncore.js"
 import { Language } from "./types.js"
 
@@ -27,7 +27,7 @@ export const deserializeLanguagesWithHandler = (
 ): Language[] =>
     deserializeSerializationChunk(
         serializationChunk,
-        lioncoreInstantiationFacade,
+        lioncoreWriter,
         [lioncore, ...dependentLanguages],
         [lioncoreBuiltins, ...dependentLanguages].flatMap(nodesExtractorUsing(lioncoreReader)),
         new BuiltinPropertyValueDeserializer(),