+++ packages/test/src/core/library.test.ts
@@ -1,15 +1,15 @@
 import {
     deserializeLanguages,
     deserializeSerializationChunk,
-    dynamicInstantiationFacade,
     DynamicNode,
+    dynamicWriter,
     lioncoreBuiltins,
     nameBasedClassifierDeducerFor,
     nodeSerializer,
     serializeLanguages
 } from "@lionweb/core"
 
-import { libraryInstantiationFacade, libraryModel, libraryReader } from "../instances/library.js"
+import { libraryModel, libraryReader, libraryWriter } from "../instances/library.js"
 import { libraryLanguage } from "../languages/library.js"
 import { deepEqual } from "../test-utils/assertions.js"
 
@@ -17,13 +17,13 @@ describe("Library test model", () => {
     it("[de-]serialize example library", () => {
         const serializationChunk = nodeSerializer(libraryReader)(libraryModel)
         // FIXME  ensure that serialization does not produce key-value pairs with value === undefined
-        const deserialization = deserializeSerializationChunk(serializationChunk, libraryInstantiationFacade, [libraryLanguage], [])
+        const deserialization = deserializeSerializationChunk(serializationChunk, libraryWriter, [libraryLanguage], [])
         deepEqual(deserialization, libraryModel)
     })
 
     it(`"dynamify" example library through serialization and deserialization using the DynamicNode facades`, () => {
         const serializationChunk = nodeSerializer(libraryReader)(libraryModel)
-        const dynamification = deserializeSerializationChunk(serializationChunk, dynamicInstantiationFacade, [libraryLanguage], [])
+        const dynamification = deserializeSerializationChunk(serializationChunk, dynamicWriter, [libraryLanguage], [])
         deepEqual(dynamification.length, 2)
         const lookup = nameBasedClassifierDeducerFor(libraryLanguage)
         deepEqual(dynamification[0].classifier, lookup("Library"))